==> Network Analysis/01. Dataset creation.ipynb <==
## Notebook where we create the Dataset with the API and Spotipy

import pandas as pd
import spotipy
from spotipy.oauth2 import SpotifyOAuth
from spotipy.oauth2 import SpotifyClientCredentials
from client import * #client credentials here
import os
import sys
import time
%matplotlib inline
import numpy as np
import networkx as nx
import matplotlib.pyplot as plt
from random import randint


SPOTIPY DOC: https://spotipy.readthedocs.io/en/2.19.0/#module-spotipy.client

# We create the client object to interact with the API using Spotipy

sp = spotipy.Spotify(auth_manager=SpotifyOAuth(client_id =Client_id,client_secret=Client_secret ,redirect_uri=Redirect_uri ))

In order to create the first set of artists, we grab a representation of artists by choice from the community we want to analize

In this case we have chosen artists from the rap/urban music spanish community. From this selection we use the API to search the collaborations which appear in their top tracks.

with the initial set and the name of the collaborators we grow again the initial list of artists as many times as we feel fit, until we have a community of which we consider representative for our analysis.

It is recommended to go back and touch upon this data in order to refine everytime a little bit more the chosen artists

# FIRST ARTIST SELECTION

# IMPORTANT!!!: the way the name is written in this list has to be exactly the way that is written in the Spotify platform

artist_list_sp = ['SFDK','Cecilio G.','Ayax y Prok','Sharif','Rels B','Bejo',
                    'Haze','Lagrimas De Sangre','Pepe : Vizio','Santa Salut',
                    'Mala Rodriguez','Luenco','Delaossa','Bizarrap','Residente','Recycled J','Natos y Waor',
                    'Hard GZ','Nikone','El Jincho','Foyone','Kase.o','Dellafuente','Morad','Cráneo',
                    'Saske','Juancho Marques','Zetazen','C. Tangana','Cruz Cafuné','Matasvandals','Toteking',
                    'Don Patricio','Ivancano','ROSALÍA','Los Chikos del Maiz','Nach','Shotta','Rapsusklei',
                    'Ptazeta','Bad Gyal','FERNANDOCOSTA','Yung Beef','Israel B','Kidd Keo']

len(artist_list_sp)

# Function that tracks the top tracks of our artists and returns the colaborators in those songs

def colab_top_tracks (artist):


  time.sleep(1)
  #     te dice el URI del artista
  busqueda_id = sp.search(q=f'artist: {artist}',limit = 1, type='artist')
  try:
      ID = busqueda_id['artists']['items'][0]['uri']
  except IndexError:
      fake_artists.append(artist)

  # Top track URIs

  results = sp.artist_top_tracks(ID) #Possibility of defining the country in order to increase the chance of finding the correct artist
  top_tracks_uris = list()

  for track in results['tracks']:
      top_tracks_uris.append(track['uri'])

  # Colabs with Top track URIs

  colabos_top_tracks = set()

  for i in top_tracks_uris:
      try:
          colabs_track =sp.track(i)
          for j in colabs_track['artists']:
              if j['name'] != artist:
                  colabos_top_tracks.add(j['name'])
              else:
                  continue
      except Indexerror:
          artists_with_no_artists.append(i)


  lk = list()

  lk.append(artist)
  lk.append(colabos_top_tracks)

  return tuple(lk)

colab_top_tracks('DELLAFUENTE')

# These lists will pick up the issues that the spotify API has on some artists, we might use it further on for our analysis
# or data clean up.

albums_with_no_songs = list()
artists_with_no_artists = list()
fake_artists = list()
artist_with_no_albums = list()

# This for loop applies the given function for the colaborators. Due to the API limits, we also included a While loop
# to keep trying as for random moments the API doesn't let us extract the desired information


list_colabs_rap = list()

for i,name in enumerate(artist_list_sp):
    print(i)
    result = None
    while result is None:
        try:
            result = colab_top_tracks(name)
        except:
            time.sleep(3)
    list_colabs_rap.append(result)

# This function collects all the colaborators and fixes them into a list

colaboraciones_ronda_1 = list()

for i in list_colabs_rap:
    for j in i:
        for x in j:
            if len(x)>1:
                colaboraciones_ronda_1.append(x)

# We eliminate duplicates from colaborators

first_round_rapers = set()

for i in colaboraciones_ronda_1:
    first_round_rapers.add(i)

len(first_round_rapers)

# We join the initial list with the colaborators

for i in artist_list_sp:
    first_round_rapers.add(i)

len(first_round_rapers)

# We do the first round of colaborators of past colaborators + the initial list

list_colabs_rap2 = list()

for i,name in enumerate(first_round_rapers):
    print(i)
    result = None
    while result is None:
        try:
            result = colab_top_tracks(name)
        except:
            time.sleep(3)
    list_colabs_rap2.append(result)

# Create the data frame
df = pd.DataFrame(list_colabs_rap2, columns=['Name', 'Colabos'])

df.head()

# export it to CSV

df.to_csv('main_database.csv', index = False)

# Creation of the adjacency matrix

for rapero in df['Name']:
    aux = list()
    for colabos in df['Colabos']:
        if rapero in colabos:
            aux.append(1)
        else:
            aux.append(0)
    df[rapero] = aux

# Cleanup of the new dataframe

df.set_index('Name', inplace=True)

df.drop('Colabos', axis=1, inplace=True)

# we export the Matrix to have it as an excel file

df.to_csv("adjacency_matrix.csv",index=True)

df.shape

df.sum().sum()

# Creation of the adjacency_list

adjacency_list = list()

for i,column in enumerate(df.columns):
    for j,fila in enumerate(df.index):
        if df.at[fila,column] == 1:
            listita = list()
            listita.append(column)
            listita.append(fila)
            adjacency_list.append(listita)
            

adjacency_list_df = pd.DataFrame(adjacency_list,columns = ['start_node','end_node'])

adjacency_list_df.to_csv("adjacency_list_noheader.csv",index=False,header=False)


adjacency_list_df.to_csv("adjacency_list.csv",index=False)

adjacency_list_df

adjacency_list_df.shape
