==> Network Analysis/03. EDA with new metrics.ipynb <==
import pandas as pd
import os
import sys
import time
%matplotlib inline
import numpy as np
import networkx as nx
import matplotlib.pyplot as plt
from random import randint
import spotipy
from spotipy.oauth2 import SpotifyOAuth
from spotipy.oauth2 import SpotifyClientCredentials
from client import * #client credentials here

# We create the client object to interact with the API using Spotipy

sp = spotipy.Spotify(auth_manager=SpotifyOAuth(client_id =Client_id,client_secret=Client_secret ,redirect_uri=Redirect_uri ))

# We bring in the main DF where we have the chosen artists

df = pd.read_csv('graph_metrics.csv')
df.head()

df.drop(['eigenvector_centrality','ID','num_of_edges'], axis= 1, inplace = True)

df.shape

# FUNCIONES PARAMETROS ARTISTAS

def funcion_popularidad(artist):

    #     te dice el URI del artista
    busqueda_id = sp.search(q=f'artist: {artist}',limit = 1, type='artist')
    try:
        ID = busqueda_id['artists']['items'][0]['uri']
    except IndexError:
        fake_artists.append(artist)
        return       
           
    #     Definimos la parametros del artista

    artist_param = sp.artist(ID)

    #    Definimos los diccionarios donde irán los parametros

    return artist_param['popularity']
    
def funcion_followers(artist):

    #     te dice el URI del artista
    busqueda_id = sp.search(q=f'artist: {artist}',limit = 1, type='artist')
    try:
        ID = busqueda_id['artists']['items'][0]['uri']
    except IndexError:
        fake_artists.append(artist)
        return       
           
    #     Definimos la parametros del artista

    artist_param = sp.artist(ID)

    #    Definimos los diccionarios donde irán los parametros

    return artist_param['followers']['total']

    
def funcion_generos(artist):

    #     te dice el URI del artista
    busqueda_id = sp.search(q=f'artist: {artist}',limit = 1, type='artist')
    try:
        ID = busqueda_id['artists']['items'][0]['uri']
    except IndexError:
        fake_artists.append(artist)
        return       
           
    #     Definimos la parametros del artista

    artist_param = sp.artist(ID)

    #    Definimos los diccionarios donde irán los parametros

    return artist_param['genres']


funcion_popularidad('Charles Ans')

# Create the popularity metric
df['popularity'] = [funcion_popularidad(i) for i in df['name']]

# Create the followers metric
df['followers'] = [funcion_followers(i) for i in df['name']]

# Create the genres metric
df['genres'] = [funcion_generos(i) for i in df['name']]

df.head()

df.to_csv('df_all_metrics.csv', index = False)

# Checking the correlation

df.corr()

top_10_popularity = (df.sort_values(by = ['popularity'], ascending = False)).head(10)
top_10_popularity

bottom_10_popularity = (df.sort_values(by = ['popularity'], ascending = True)).head(10)
bottom_10_popularity
