==> Network Analysis/04. Tools & Conclusions.ipynb <==
# Tools & Conclusions

import pandas as pd
import os
import sys
%matplotlib inline
import numpy as np
import networkx as nx
import matplotlib.pyplot as plt
from random import randint
import pickle

First tool corresponds to an analysis of the network. As we have already done an analysis the network as a whole, we are now going to analyse a representative community

df = pd.read_csv('df_metrics_communities.csv')

# We clearly pick mainstream_latino

df['communities'].value_counts()

df_latino = df[df['communities'] == 'mainstream_latino']

df_latino.head()

Now we are going to analyse all the measures 1 by 1

df_latino.sort_values(by = ['degree_centrality'], ascending = False).head(8)

df_latino.sort_values(by = ['betweenness_centrality'], ascending = False).head(8)

df_latino.sort_values(by = ['closeness_centrality'], ascending = False).head(8)

df_latino.sort_values(by = ['page_rank'], ascending = False).head(8)

Conclusions:

We can observe clearly that the spotlight nodes are:

Bizarrap, Nacho, Rels B & Bad Bunny

This makes sense as Bizarrap is a DJ/Music producer and right now has contact with all of this artists, so collaborating with him has great repercussion. It has been seen when one of his las collaborations with the artist Residente received over 60 Million views in 3 days.

Regarding the other spotlight nodes, Nacho & Bad bunny are Reguetton artists with many collaborations with other nodes that are in most of our top 8 like Duki and are also nodes that have a high response towards the network.

In this analysis however, Rels B is a spotlight node which caught my interest as they are Spanish artists and have moved over from Spain more towards the latino community, with their style and connections within that community. They are a great example and connection from those other artists that would want to move into the latino_mainstream community!


# Shortest paths

Following our example we will grab an artist from a low populated community and make a route to the latino_mainstream

# We clearly pick trap lofi

df['communities'].value_counts()

df_trap_lofi = df[df['communities'] == 'trap_lofi']

# lets grab a not that well connected artist and help him/her out to the latino community by using PageRank

a = df_trap_lofi.sort_values(by = ['page_rank'], ascending = True).head(8)
a[['name','page_rank']]

Now lets analyse the routes to Rels B


adjacency_list_df = pd.read_csv('adjacency_list.csv')

# Graph creation:

G = nx.from_pandas_edgelist(adjacency_list_df, 'start_node', 'end_node')

# computing the shortest path length for all pairs of nodes in the network:
# here we attain all the artists and the number of nodes that they have to go through to get to certain parts of the network

shortest_path_lengths = dict(nx.all_pairs_shortest_path_length(G))

# from the most popular artists in this community LUNA KI is far enough

shortest_path_lengths['LUNA KI']['Rels B']

# These are all the possible paths of the network 
# The more nodes in the middle to reach to an artist the more possible routes there are


paths = nx.all_simple_paths(G, source='LUNA KI', target='Rels B', cutoff= 8)
paths_posibilities = list(paths)

len(paths_posibilities)

df_communities = pd.read_csv('df_metrics_communities.csv')

df_communities.columns

# We are creating a Dataframe with all the possible routes and 
# within them the means and std dev of their respective popularities

all_paths_std_mean = list()

for route in paths_posibilities:
    artist_route = list()
    for artist in route:
        artist_route.append(df_communities[df_communities['name'] == artist]['popularity'].values[0])
    all_paths_std_mean.append((tuple(route),np.array(artist_route).mean(),np.array(artist_route).std()))


df_best_route = pd.DataFrame(all_paths_std_mean, columns = ['route', 'mean', 'standard_dev'])

In order to reduce the number of paths we will work with the highest popularity mean and lowest std deviation for each route to find the most popular artists which are all at the same level. This will increase the likelihood of a greater impact with the artist that we are trying to reach.


df_best_route.describe()

# For this case, the following routes would be the optimal ones. 
# From here we should consider with the artists which are
# the best approaches

LUNA_KI_best_route = df_best_route[(df_best_route['mean'] > 76) & (df_best_route['standard_dev'] < 9.5)]
LUNA_KI_best_route

These would be the best routes to reach Rels b

LUNA_KI_best_route['route'].values
